+++ tests/testthat/test-joins.r
@@ -422,4 +422,19 @@ test_that("inner join gives same result as merge. #1281", {
   y <- data.frame(cat1 = sample(c("A", "B", NA), 5, 1),
     cat2 = sample(c(1, 2, NA), 5, 1), v = rpois(5, 3),
     stringsAsFactors = FALSE)
+  ij <- inner_join(x, y, by = c("cat1","cat2"))
+  me <- merge(x, y, by = c("cat1","cat2"))
+  expect_true( equal_data_frame(ij, me) )
+})
+
+test_that("join handles matrices #1230", {
+  df1 <- data_frame( x = 1:10, text = letters[1:10])
+  df2 <- data_frame( x = 1:5,  text = "" )
+  df2$text <- matrix( LETTERS[1:10], nrow=5)
+
+  res <- left_join( df1, df2, by = c("x"="x")) %>% filter( x > 5)
+  text.y <- res$text.y
+  expect_true( is.matrix(text.y) )
+  expect_equal( dim(text.y), c(5,2) )
+  expect_true( all(is.na(text.y)) )
 })